+++ ConfigDrawers/Properties/AssemblyInfo.cs
@@ -2,7 +2,7 @@ using System.Reflection;
 using System.Runtime.InteropServices;
 
 [assembly: AssemblyTitle("BepInEx.ConfigDrawers")]
-[assembly: AssemblyDescription("Next-generation in-game Configuration Manager for Unity BepInEx")]
+[assembly: AssemblyDescription("In-game Configuration Manager for Unity BepInEx")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("Vapok")]
 [assembly: AssemblyProduct("BepInEx.ConfigDrawers")]
@@ -13,5 +13,5 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 [assembly: Guid("5E4029F7-70DB-4C5C-BDC3-A882E9437DEB")]
 
-[assembly: AssemblyVersion("1.0.2.0")]
-[assembly: AssemblyFileVersion("1.0.2.0")]
+[assembly: AssemblyVersion("1.1.0.0")]
+[assembly: AssemblyFileVersion("1.1.0.0")]